+++ pom.xml
@@ -1,10 +1,10 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
     <modelVersion>4.0.0</modelVersion>
-    <groupId>duowan</groupId>
-    <artifactId>sql</artifactId>
+    <groupId>github.jadetang</groupId>
+    <artifactId>scala_sql</artifactId>
     <version>1.0-SNAPSHOT</version>
-    <inceptionYear>2008</inceptionYear>
+    <inceptionYear>2015</inceptionYear>
     <properties>
         <scala.version>2.10.4</scala.version>
     </properties>
